Increase maximum command arguments from 127 to 4095
Using bit fields, to take 4 bits from Command::min_args and give
it to Command::max_args, without increasing the size of the struct
(to avoid bloating the static array of commands).

+++ src/parse-args.h
@@ -1,6 +1,6 @@
 #ifndef PARSE_ARGS_H
 #define PARSE_ARGS_H
 
-const char *parse_args(char **args, const char *flag_desc, int min, int max);
+const char *parse_args(char **args, const char *flag_desc, size_t min, size_t max);
 
 #endif
